// src/app/interfaces/column-attributes.interface.ts
export interface IColumnAttributes {
  editType: string;
  field: string;
  isPrimaryKey: boolean;
  headerText: string;
  format: string;
  width: string;
  textAlign: string;
  fontSize: number;
  color: string;
  defaultValue: any;
  minWidth: number;
  backgroundColor: string;
  textWrap: string;
}